backend/requirements.txt
# ========================================
# BACKEND DEPENDENCIES
# ========================================

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# OpenAI
openai==1.10.0

# HTTP Client
httpx==0.26.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Logging
python-json-logger==2.0.7

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-slugify==8.0.1

# Production
gunicorn==21.2.0
